~ move spec to record where it fits better
